Parse gps.conf and invoke pdsm_xtra_set_auto_download_params() only once until reboot.

+++ leo-gps.c
@@ -82,7 +82,7 @@ void update_gps_status(GpsStatusValue value);
 void update_gps_svstatus(GpsSvStatus *svstatus);
 void update_gps_nmea(GpsUtcTime timestamp, const char* nmea, int length);
 
-extern int get_cleanup_value();
+extern uint8_t get_cleanup_value();
 
 /*****************************************************************/
 /*****************************************************************/
